Add a run hook to `ActiveRecord::Migration`
In order to support our deployment process, we'd like to include a
module in `ActiveRecord::Migration`.  We don't want to load the
class unnecessarily.  `ActiveSupport` provides a method to run code
when a class is loaded, but it depends on a hook having been
defined for the class in question.  This commit adds such a hook
when the Migration class, and its support classes in the same file,
have been loaded.