Fix #54805.
When using postgres adapter and the `structure_load_flags` options, the
extra flags were appended instead of prepended to the default ones,
causing the `psql` command to ignore some of the extra flags.

Now, the default arguments `args` are appended to the `extra_flags`
instead of the opposite.

Example of `extra_flags` that were ignored by postgres:

    ActiveRecord::Tasks::DatabaseTasks.structure_load_flags = ['-c', 'SET my_extension.level=off;'].freeze

Now, they will be taken into account.