Deprecate ActionController.escape_json_responses= at the method
Instead of emitting the deprecation in an initializer, we should emit
the deprecation when calling the writer method.

In order to reduce warnings when users are updating their apps to use
the new default `true`, we check the value before emitting.

This includes the class attribute method as well.

In order to set the value for the internal attribute, we need to find
out the name of the internal setter.

https://github.com/rails/rails/blob/1c094d762d25805cc2d110ee2f7b54b33352e293/activesupport/lib/active_support/core_ext/class/attribute.rb#L95-L102

In the future, it would be nice if `class_attribute` and `mattr_*`, etc
had a way to easily deprecate parts of the methods being defined, or
provide a callback (so we can emit warnings based on the value, etc).

Co-authored-by: Alex Ghiculescu <alex@tanda.co>
Co-authored-by: Jean Boussier <jean.boussier@gmail.com>