Copy Active Storage test fixture racecar.jpg into Railties fixtures
This change is specifically to support packaging vendors (Fedora) who
test gems in isolation, they don't have the source of the whole Rails
checkout available, and the test files are not packaged with the
`.gem`s. so this is a comprimise so they can still execute this test in
their environment.