Move AbstractAdapter#valid_type? to a class method
since this method only needs access to a constant, and does not
actually need a live connection.

Continue to support instance methods `#valid_type?` and
`#native_database_types`, but both of these methods now forward to a
class method of the same name. (All the concrete adapters now
implement a class method `.native_database_types`.)

This allows code without a connection to check the column type without
leasing a connection.