Add `default_parameter_encoding` method for specifying default parameter encodings in a given action.
Developers can use the `param_encoding` method to specify the character encoding for a parameter.
However, it requires a one line for each parameter.
This is tiresome for developers and may also lead to bugs if they forget to specify the encoding.

This PR introduces the `default_parameter_encoding` method to simplify this.
`default_parameter_encoding` allows specifying a default encoding for all parameters in a given action.

This enhancement makes the codebase of applications receiving non-UTF-8 encoded data cleaner and easier to maintain.