ActiveStorage::Streaming limit range requests to a single range
Multi-Range requests can easily be abused DoS the service.

The backend could and should be optimized to handle multi-range
requests better, but ultimately there isn't a whole lot of legitimate
uses for them, and if you are exposing Active Storage to your
users with large files, you should consider not using the default
proxy.

[CVE-2026-33658]
[GHSA-p9fm-f462-ggrg]