Add ActiveRecord::Base.establish_connection behavior change to Rails 7.2 release notes.
While part of our upgrade from Rails 7.1 to Rails 7.2, one of the issues we encountered, and that was not very straightforward to find, was that things broke because we assumed that `ActiveRecord::Base.establish_connection` would make the connection active.

Then we found this issue https://github.com/rails/rails/issues/52858 which helped us solve the issue.

If this change had been part of the Rails 7.2 release notes, it would have saved some time.