Use `TRUE` and `FALSE` for booleans in SQLite
Previously, queries generated with boolean comparisons for MySQL and
Postgres used `TRUE` and `FALSE` (these are actual booleans in Postgres
and aliases for `1`/`0` in MySQL) but used `1` and `0` for SQLite. This
makes it more difficult to write database-agnostic query assertions,
since the values are different.

SQLite 3.23.0 added support for `TRUE` and `FALSE` as aliases for
`1`/`0` (similar to MySQL). A [previous commit][1] bumped the minimum
required SQLite version in Active Record to 3.23.0, and this commit
updates the boolean query generation for SQLite to use `TRUE` and
`FALSE`.

[1]: 809abd3ed3c7700bae1edf104c0ad61acb638d4b