Avoid unnecessary splat allocations
In these 5 calls, an array is allocated for the splat, as the Ruby
compiler cannot determine that the keyword/block expressions will
not have side effects that will affect the splat. By assigning the
related expression to a local variable and using the local variable,
you can avoid the allocation for the splat.

Found by the performance warning in Ruby feature 21274.