Enable passing retryable `SqlLiteral`s to `#where`
Previously, retryable SqlLiterals passed to `#where` would lose their
retryability because both `#build_where_clause` and `WhereClause` would
wrap them in non-retryable `SqlLiteral`s.

To fix this problem, this commit updates `#build_where_clause` to check
for `SqlLiterals` and updates `WhereClause` to assume that any predicate
passed in will already be wrapped.

There were only a few places that passed Strings to `WhereClause`
(`"1=0"` in `PredicateBuilder` and sanitized sql strings in
`#build_where_clause`) that needed to be updated to use `Arel.sql`. The
`WhereClause` tests also had Strings to wrap but they are really unit
tests and aren't representative of what can be done with the public API