action_cable_meta_tag uses script name to build the mount path
When Rack::Request#script_name is being used, the Action Cable meta
tag should be setting a mount path that includes the script name.

This PR adds a name to the Action Cable server route, so that it can
be referenced from the meta tag helper, where the script name is injected.