Use TRUE and FALSE for more SQLite queries
A [previous commit][0] updated the SQLite Arel visitor to use `TRUE` and
`FALSE` literals instead of `1` and `0` for `True` and `False` Arel
Nodes.

This commit takes the change further by making SQLite quote `true` and
`false` values as `TRUE` and `FALSE` instead of `1` and `0`. This
additionally required adding support for `DEFAULT TRUE` and `DEFAULT
FALSE` SQLite column definitions (which is more correct anyways in case
someone made `TRUE` or `FALSE` the default value of a column using
`default: -> { "TRUE" }`, etc).

[0]: 34bebf383e18243a1cdadc461e3a84c66125cb9b