Avoid DISTINCT on COUNT queries when possible
* Fixes #55027

Previously, all queries with eager-loaded associations resulted in  queries.
However, this addition of DISTINCT is unnecessary and can be expensive on
larger tables when the base table belongs_to all eager-loaded associations.

This change adds a check for any referenced or eager-loaded tables on
calculations and skips the DISTINCT addition if all joined tables ar
belongs_to associations.