Fix affected_rows for SQLite sql.active_record notifications
SQLite's connection.changes only reflects the number of rows affected by the last
DELETE/UPDATE/INSERT, and the number would not change for any other query. Therefore,
SELECT statements would have the affected row count of a previous query. To solve this,
we are now calculating the different between the total changes before and after the
query has been executed.

Co-authored-by: Hartley McGuire <skipkayhil@gmail.com>