Add cancel button to first version of the new Product page
In the section "Extracting Partials", a cancel button suddenly appears in `app/views/products/new.html.erb`, which slightly distracts from the focus on how to extract a partial view.

Putting the cancel button in the first version of the new product view reduces the number of new things in the Extract Partials code snippet.