Revert "Deprecate :class_name for polymorphic belongs_to"
In order to upgrade to a minor/major version of Rails, I
thought we ask people to upgrade to the last patch release
in their current version and *then* fix deprecations to
make sure they have them all.

That is, I believed that was explained that way because we
could add deprecations in patch releases to be able to make
breaking changes in the next minor/major.

But I was mistaken, once a stable release is out, the
established practice is to not add any further deprecations.

Thanks to @Earlopain and @matthewd.

This reverts commit a1ba71b5150cc37b556bfea6ddab23db81459df9.