Set cache `max-age` values consistently in configs
Set cache `max-age` values consistently in configs
so there is no difference in the requirements for the environment
config files and it is easier to compare different environments.

At one point it was accepted that these values
should use ActiveSupport methods to make the value self-explanatory:
https://github.com/rails/rails/commit/33f116f22cd46c26ecc60e7f9f9d2bd0691b9fbb

However, more recently this was reverted in the `test` environment
with the rationale that it wasn't worth the ActiveSupport require:
https://github.com/rails/rails/commit/44548856c4c50e23bae0d007d90bfaf95b87d1d7

This commit updates the other environment config files to match `test`
as I don't see a reason why they should be different.