Always fully clear current attributes
Attempting to reset instances is a recipe for causing state leak.

Regular instance variables in `Current` subclasses aren't cleared,
so unless you remember to undefine them in the `reset` callback
they will leak across requests.

We could of course document that, but it's just too much of a footgun,
and I see no benefit of trying to re-use that object instance.

It's much safer to just always drop it and start fresh.

Closes: https://github.com/rails/rails/pull/55125/

Co-Authored-By: Janko Marohnić <janko@hey.com>