Add ability to change the transaction isolation for all pools within a block
rails/rails#54836 introduced the ability to change the isolation for the
current pool. If you're changing the default transaction isolation for an
app with many dbs/shards you would need to loop through all pools and
apply the isolation. If there are hundreds of connections, that might
create a performance issue.

This change adds a new method
`ActiveRecord.with_transaction_isolation_level` which will change
the transaction isolation for any pool accessed within the block. The
method can be used in an around filter to update all transactions for an
action or middleware to switch all pools accessed in a request.

I chose to call this from `with_transaction_returning_status` and
`transaction` because called from within `connection.transaction` made
it too difficult to tell if we were changing the isolation on an open
transaction. It would either error when it wasn't actually changing or
fail to error when it was supposed to. Wrapping the calls solves this.

While implementing this I felt that the original method had a confusing
name paired with the new one so I updated it to be more explicit that
it's for only the currently current pool.