Enable configuring Action View's render tracker
In Rails 7, a new render tracking implementation was [added][1] to
Rails: the RipperTracker. From that commit message:

> Using a parser finds dependencies that would otherwise be difficult to
> find with the regular expressions. It should also theoretically work
> with other template systems since it operates on the compiled template
> instead of the contents of the file.

It was initially made the default render tracker, but that change was
later [backed out][2].

Since then, the Prism parser was released and made the default Ruby
parser. A Prism render tracker implementation was also [added][3] to
Rails in 7.2.

Until now, there hasn't been a "blessed" way for applications to try out
the new tracker implementations (you would have to go through private
API). This commit adds configuration to not only make it easy for
applications to use the new Ruby-parser-based trackers, but it also
makes these new trackers the default for new applications on 8.1.

The new trackers are not only better at finding `render` calls, they are
also much easier to maintain. When I [added][4] trailing interpolation
tracking to the three trackers, the Prism tracker was by far the easiest
to work with.

[1]: 897b9bf6e29a9aa7c04d3e4cb96a65665565154d
[2]: 41c0681fb7626d036005f8dec559a507ff143874
[3]: d743a2b10fb50dc4702ba4e73955c4f94b3bf700
[4]: 13df150c06e37a7024baa655000ae2d7eff9fc0c