Fix `include?` on strict-loaded HABTM association
Previously, trying to check whether a `new_record?` (`Post.new`) is
`include?`d in a strict-loaded HABTM association would raise a
`StrictLoadingViolationError` even when the HABTM association is already
loaded.

This commit fixes the issue by always using the `target` if the
association is `loaded?` instead of querying the Through association
(which is only needed when the association is dirty/has newly built
records).