Active Job Continuation isolated steps
Add an isolated option to steps. Defaults to false.

Isolated steps are always run in their own job execution. We do this by
interrupting the job before running a step if both:

1. It is not the first step we've run in this execution
2. Either this step or the previous one was marked as isolated.

The first to be run in an execution is always run inline so we don't
end up in an infinite loop of interruptions.

This allows you to execute a long running step separately which is
useful to ensure that progress is saved before it runs.