`has_secure_password`: fix password validation.
Previously, the password confirmation validation added by
`has_secure_password` was skipped if the password string was a sequence
of whitespace characters, regardless of the string provided as
confirmation, which was ignored with no error messages.

This change fixes that behavior, so that the confirmation validation
runs consistently, independently of the password string content.

Fixes #55225