Make sure that `CurrentAttributes.set` works on the current instance
If the block passed to `CurrentAttributes.set` happens to clear all
current attributes (which happens for example when performing jobs),
then `.set` is unable to restore attribute values on the proper
singleton instance of the `CurrentAttributes` class.
See the added test for details.

The problem has always existed, but was emphasized by PR #55139, causing
tests that perform enqueued jobs in a `Current.set` block to reset
current attributes to their defaults when exiting the block,
instead of the values that were set previously in the test.