Allow for nested ExecutionContext in test
`ActiveSupport::CurrentAttribute` lifetime in test always have been
a bit wonky.

This was addressed in the past by https://github.com/rails/rails/pull/43734
and then https://github.com/rails/rails/pull/43550 but a problem remained
that various states would leak from the test case into the controller
or job being tested, and back.

Ideally when executing a controller or job from a test case, we
want to "save and restore" current attributes and the execution context.

This is what this commit does, by implementing `push` and `pop` logic
for `ActiveSupport::ExecutionContext`.

For extra safety, since such use case isn't supposed to ever happen
in production, this behavior is test only and in production we keep
dropping all the state.