[RF-DOCS] Rails Application Template Guide - merge with Rails Generators Guide [ci-skip] (#55020) (#55261)
The goal of this PR is to update the Rails Application Template Guide. In the process of updating this guide, it made sense to merge its content into the Generators Guide, which already had a section about Templates. The scope of this PR is the Rails Application Template sections (the rest of the existing Generators Guide will be updated later on).

Co-authored-by: bhumi1102 <bhumi1102@gmail.com>
Co-authored-by: Ridhwana <Ridhwana.Khan16@gmail.com>