Add .rubocop_todo.yml to RuboCop cache of GHA workflow template
Follow-up to https://github.com/rails/rails/pull/54754.

At the time of running `rails new`, .rubocop_todo.yml does not exist.
This file is automatically generated by commands such as `rubocop --auto-gen-config`.
If changes in caching behavior occur after such a command generates .rubocop_todo.yml,
it may be difficult to notice the impact.

With this change, users no longer need to worry about the presence of .rubocop_todo.yml,
which is frequently used with RuboCop, in relation to the cache in GHA workflows.