Fix key collisions in HashWithIndifferentAccess#transform_keys
Because the current implementation of transform_keys happens while
iterating instead of atomically, transformations that result in
collisions with pre-transformed keys can replace subsequent key/value
pairs.

For example where Hash will give you this:

    sym_hash = { a: "one", aa: "two", aaa: "three", aaaa: "four" }
    sym_hash.transform_keys { |k| "#{k}a".to_sym }
      #=> {:aa=>"one", :aaa=>"two", :aaaa=>"three", :aaaaa=>"four"}

HashWithIndifferentAccess was returning this:

    sym_hwia.transform_keys { |k| "#{k}a".to_sym }
      #=> {"aaaaa"=>"one"}

This fix ensures that all keys are transformed as a collection.