Make `ActiveSupport::Gzip.compress` deterministic
`ActiveSupport::Gzip.compress` used to include the current timestamp in the gzipped output (the default behavior of Zlib), causing multiple calls with the same input data to output different results depending on when it was called. In many situations (such as caching), ensuring that the output is always the same for a given input is useful, so this helper now always sets the timestamp to `0` instead.