Fix kill_connection_from_server for PG test
and also fix incorrect comment for different test method

This method is intended to make a subsequent query hit the rescue inside
`with_raw_connection`, but this was only actually happening with
`mysql2` and `trilogy`. The `pg` adapter was not being correctly put
into a state where the `SELECT` in the test would need to retry.

This commit fixes the issue by using pg_terminate_backend instead of
pg_cancel_backend. Now when the test is run with `pg` it correctly hits
the rescue in `with_raw_connection`.