Log when rescue_from handles an exception
In development when a `before_action` renders or redirects this error
message appears on the logs:

> Filter chain halted as :before_action rendered or redirected

I find this message very helpful. However, if a before_action raises an
error that is handled by a rescue_from block, nothing is logged
indicating an error occurred.

This PR adds a log message indicating an exception was rescued, and
displays where it was originally raised. For example:

> rescue_from handled BaseController::YourError (Oh no!) -
>   app/controllers/home_controller.rb:9:in 'HomeController#authenticate'

Co-Authored-By: Jean Boussier <jean.boussier@gmail.com>