Allow to use either `:unprocessable_entity` or `:unprocessable_content`
There is just so much code out there referencing the old status
that it's unreasoable to ask Rails users to migrate, especially
since there is 0 gain from the rename.

It's even more problematic because there are many gems and such,
so they all need to be updated to check the Rack version to know
which symbol to emit.