Allow local_cache_store_strategy in initializer
Previously, the `Rails.cache.middleware` was added to the stack in a
bootstrap initializer, which meant trying to disable it via a
configuration in an initializer wouldn't work (althrough it could still
be removed via middleware stack operations).

This commit moves `Rails.cache.middleware` setup to a later point in
initialization, middleware stack building, which then makes it possible
to disable its inclusion via normal configuration in an initializer.