Enable hooking LocalCache into Executor
Previously, Rails would automatically add the `LocalCache::Middleware`
to the stack if `Rails.cache` is configured to use a `Store` that
supports `LocalCache` (currently, `MemCacheStore`, `RedisCacheStore`,
and `NullStore`).

This commit introduces an additional option to the configuration added
in a previous commit (`config.local_cache_store_strategy`), `:executor`.
When using this option, Rails will register `run` and `complete` hooks
on the application's executor instead of using a middleware.

While Rails automatically adds a middleware for requests to take
advantage of the `LocalCache`, there is no such built in mechanism for
Active Job. Similarly, regular `ActiveSupport::TestCase` tests will not
end up using the `LocalCache` even though they may be testing behavior
that _would_ go through the `LocalCache` during a real request. By using
the executor for configuring the `LocalCache`, requests, job, tests, and
any other application code wrapped in the executor will be able to take
advantage of it. This is also already the behavior of Active Record's
in-memory-cache, so this would align `Rails.cache`'s in-memory-cache
with it.

Finally, using the executor means we have one less middleware in the
request stack and one less BodyProxy allocated and wrapped around the
response, which makes me happy.