[rail_inspector] Fix linting defaults post-release
Previously, the linter would not validate that all versions with
defaults were documented, just that documented versions were correct.
This lead to no validation of defaults for a new version until the new
version's header was added.

Additionally, the header couldn't be added until the first default for
that version was added because the linter would raise an error if a
version header was present with no defaults.

This commit fixes both of those problems. It now more strictly compares
the _actual_ versioned defaults with the documented defaults so that it
will never miss any based on which are currently documented.
Additionally, it won't error if a header is present without any defaults
(and due to the more strict comparison, this state is expected).