Extract (nestable) Execution from ExecutionContext
This extraction separates the two things that the ExecutionContext deals
with: per-request/job/etc. Context and the ability to restore
per-request/job/etc. state when nested in test environments.

This makes it simpler to give additional state the ability to work with
nested Executors in the test environment. This commit demonstrates what
that change looks like for the two states (Context and
CurrentAttributes).