Allow logs and notifications for open redirects
Instead of only allowing to raise.

This is to aid large codebases to find all places that currently do open
redirects but are not called with `allow_other_host: true`.

`raise_on_open_redirects` is marked as deprecated and if set to true
will set `action_on_open_redirect` to `:raise`