Fix truncate clearing query cache twice
Since truncate calls execute and both of those methods are intercepted
to clear the query cache, it unecessarily clears the cache twice.

This commit removes the truncate delegation so that the cache only gets
cleared once (with execute). A test is added to ensure that this doesn't
regress if the implementation of truncate were to be changed in the
future.