Remove TableMetadata -> Reflection delegation
The delegation is unnecessary since the reflection instance is already
available where needed in the predicate builder.

Removing it simplifies TableMetadata (as well as the classes it was
previously passed to only to delegate to a reflection), and prevents a
TableMetadata allocation except when actually used (querying a through
association).