Turn off debug mode when calling assert_no_event_reported
rails/rails@5cc95ee made debug mode the default in test. However, when
calling assert_no_event_reported developers likely want to test
production behaviour, so we should turn off debug mode when calling
this assertion.

We still give the user the option to keep it on with the :debug param.